net: dsa: tag_8021q: change order of rx_vid setup
The 802.1Q tagging performs an inbalanced setup in terms of RX VIDs on
the CPU port. For the ingress path of a 802.1Q switch to work, the RX
VID of a port needs to be seen as tagged egress on the CPU port. While
configuring the other front-panel ports to be part of this VID, for
bridge scenarios, the untagged flag is applied even on the CPU port in
dsa_switch_vlan_add.

Fix this by moving the setup part of the RX VID on the CPU port after
all the other front-panel ports have been added to the VID.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>
Fixes: f9bbe4477c30 ("net: dsa: Optional VLAN-based port separation for switches without tagging")