net: phylink: Make fixed link notifier calls edge-triggered
Otherwise .mac_link_up and .mac_link_down on fixed-link interfaces are
called every second.

On a raw phylink instance we cannot call netdev_carrier_ok for an
indication of the last state of the link, so keep track and update this
in a new field of the phylink structure.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>