enetc: WA for MDIO register access issue
Due to a hardware issue, an access to MDIO registers that is concurrent
with other ENETC register accesses may lead to the MDIO access being dropped
or corrupted.  The workaround introduces locking for all register accesses
to the ENETC register space.  To reduce performance impact, a readers/
writers locking scheme has been implemented.  The writer in this case is
the MDIO access code (irrelevant whether that MDIO access is a register
read or write), and the reader is any access code to non-MDIO ENETC
registers.  Also, the datapath functions acquire the read lock fewer times
and use _hot accessors.  All the rest of the code uses the _wa accessors
which lock every time a register is accessed.

Signed-off-by: Alex Marginean <alexandru.marginean@nxp.com>
Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Claudiu Manoil <claudiu.manoil@nxp.com>