bus: fsl-mc: fix dprc object reading race
This is an alternative implementation to the /sys/bus/fsl-mc/rescan
hack to fix the DPRC object reading race.

Fix the race reading the current set of dprc objects by checking the
number of objects after reading them is the same as the number of
objects before reading.

This solves the issue where network interfaces disappear while adding
others via ls-addni.

This does *not* solve the issue that if an object is deleted while
another is added while we are reading the objects - that requires
firmware modification, or a more elaborate solution on the Linux side
(hashing the objects and reading all objects at least twice.)

However, without firmware modification, this is probably the best way
to ensure that we read all the objects.

Signed-off-by: Russell King <rmk+kernel@armlinux.org.uk>